+++ src/GNR1_Firmware/rev_E/NRF/MD_MASTER/FTEX_MODULES/Src/vc_interface.c
@@ -31,7 +31,7 @@ int32_t VCI_ReadRegister(VCI_Handle_t* pHandle, uint16_t RegID)
 }
 
 //TODO: Complete register table
-void VCI_SetRegister(VCI_Handle_t* pHandle, uint16_t RegID, uint8_t value)
+void VCI_SetRegister(VCI_Handle_t* pHandle, uint16_t RegID, int32_t value)
 {
 	switch (RegID)
 	{
